+++ lib/gds_api.rb
@@ -1,4 +1,6 @@
+require "addressable"
 require "plek"
+require "time"
 require "gds_api/asset_manager"
 require "gds_api/calendars"
 require "gds_api/content_store"